Set error indicator if a cue will assign an input to a DCA when it is already assigned elsewhere.
The cue-level model/view prevents a user from assigning an input to two (or more) DCAs at once.

However, when using the ListLayout, it is still possible to assign an input to two DCAs at once:
1. Assign an input to a DCA in a given cue.
2. In an earlier cue, assign input to a different DCA.

As such, this only affects the ListLayout (which uses the mapper), and has no affect on the CartLayout.
 - The CartLayout has a similar, but different, issue that I might address at a later date.