Cleanly disable the Tracker view when in Cart mode
It would be nice to keep it, however the CartLayout gui doesn't permit injecting
additional gui.

Specifically, it uses a CartTabWidget as its view, which only allows for the
stack of tabs (pages) and the tab-bar. No space for any additional gui there.

Each tab/page is then split into the grid that cue widgets are placed upon.
Trying to insert gui there is liable to interfere with cue widget access. Not
saying it's not possible, but something to look at another day.

Anyhow, this prevents LiSP throwing an error because the Tracker's view no longer
exists in the gui. It was still accessible as a member of the DcaPlotter class,
but trying to do something with it was not possible.