Support Mic/Fx and desk channel numbers not being the same
e.g. Microphone 14 is actually Channel 18 on the desk.

The UI for this in the Session Config was already in place, as support saving
channel numbers to the show file. All was missing was code to actually use it.

The conversion between assignment number and actual desk channel is done JIT. This
is to keep things simple when creating and managing assignment blocks in the Tracker
model.