Add beginning of new UI for Role Assignment
This marks the start of implementing a new "Roles" concept.

The idea behind this is thus: if an actor is double mic'd, or has an understudy
on a different mic, then this feature will eventually allow assigning those mics
to a single "Role".

Thus, should something happen mid-show, then with a button press (in a UI still
to implement), the secondary (or understudy's)  mic is used instead of the
primary mic in future DCA (re)assignments.