Support edgecases when two roles have same assignment
In the case where two Roles had the same Input (or FX) assigned to it, and we
unassign one Role and assign the other in the same Cue, it would only work if
the second Role was assigned to a DCA to the right of the first.

Assigning right to left, or replacing in place, now works.